Revert "refactor[api-server]: auth 패키지 상위 패키지로 위치 변경"
This reverts commit 313f2a4bf0dd461d9f17106586a04f707766df98.
