+++ src/main/java/com/watermelon/server/firebase/FirebaseTokenVerifier.java
@@ -3,8 +3,8 @@ package com.watermelon.server.firebase;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseAuthException;
 import com.google.firebase.auth.FirebaseToken;
-import com.watermelon.server.auth.exception.InvalidTokenException;
-import com.watermelon.server.auth.service.TokenVerifier;
+import com.watermelon.server.randomevent.auth.exception.InvalidTokenException;
+import com.watermelon.server.randomevent.auth.service.TokenVerifier;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
