+++ src/main/java/com/watermelon/server/randomevent/auth/utils/AuthUtils.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.auth.utils;
+package com.watermelon.server.randomevent.auth.utils;
 
 public class AuthUtils {
