+++ src/main/java/com/watermelon/server/randomevent/link/controller/LinkController.java
@@ -1,8 +1,9 @@
 package com.watermelon.server.randomevent.link.controller;
 
-import com.watermelon.server.auth.annotations.Uid;
+import com.watermelon.server.randomevent.auth.annotations.Uid;
 import com.watermelon.server.randomevent.link.dto.MyLinkDto;
 import com.watermelon.server.randomevent.link.service.LinkService;
+import com.watermelon.server.randomevent.link.utils.LinkUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -10,6 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import static com.watermelon.server.randomevent.link.utils.LinkUtils.makeUrl;
